refactor(wdl-lint)!: promote `CommandSectionIndentation` to analysis
