+++ crates/wdl-analysis/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   and `Document::imported_workflow_by_name` are now public
   ([#999](https://github.com/stjude-rust-labs/sprocket/pull/999)).
 * `ExceptDirectiveValid` rule, which ensures that `#@ except` comments are placed in valid locations ([#1125](https://github.com/stjude-rust-labs/sprocket/pull/1125)).
+`CommandSectionIndentation` rule, which ensures that command sections are indented consistently ([#1144](https://github.com/stjude-rust-labs/sprocket/pull/1144)).
 
 #### Changed
