+++ crates/wdl-analysis/tests/analysis/command-mixed-same-line/source.wdl
@@ -1,5 +1,3 @@
-#@ except: MetaDescription, ExpectedRuntimeKeys, EmptyOutputs, BashSetSyntax
-
 ## This is a test of having mixed indentation on the same line in command sections.
 
 version 1.1
@@ -23,7 +21,6 @@ task test2 {
 
     parameter_meta {}
 
-    #@ except: HereDocCommands
     command {
     		this line starts with spaces and ends with tabs
     }
