+++ crates/wdl-analysis/tests/analysis/command-mixed-trailing/source.wdl
@@ -1,5 +1,3 @@
-#@ except: MetaDescription, ExpectedRuntimeKeys, ShellCheck, DeprecatedPlaceholder, EmptyOutputs, BashSetSyntax
-
 ## This is a test of having mixed _trailing_ indentation in command sections.
 ## There should be no warnings from the `CommandSectionIndentation` rule.
 
@@ -24,7 +22,6 @@ task test2 {
 
     parameter_meta {}
 
-    #@ except: HereDocCommands
     command {
         this line is prefixed with ${"spaces and has tailing mixed indentation"}  		
     }
