+++ crates/wdl-lint/CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Removed
 
 * The `ExceptDirectiveValid` rule was promoted to `wdl-analysis` ([#1125](https://github.com/stjude-rust-labs/sprocket/pull/1125)).
+* The `CommandSectionIndentation` rule was promoted to `wdl-analysis` ([#1144](https://github.com/stjude-rust-labs/sprocket/pull/1144)).
 
 ## 0.26.0 - 2026-08-05
