+++ crates/wdl-lint/RULES.md
@@ -8,7 +8,6 @@ This table documents all `sprocket` lint rules implemented on the `main` branch 
 |:----------------------------|:------------------------------------------------------------|:--------------------------------------------------------------------------------------------------------------------------------------------|:--------------------------------------------------------|
 | `BashSetSyntax`             | Correctness                                                 | Ensures consistent [bash set](https://www.gnu.org/software/bash/manual/html_node/The-Set-Builtin.html) commands in every `command` section. | * `bash_set_options` - A list of required `set` options. |
 | `CallInputKeyword`          | Deprecated, Style                                           | Ensures that the `input:` keyword is not used in call statements when WDL version is 1.2 or later.                                          |                                                         |
-| `CommandSectionIndentation` | Spacing, Clarity, Correctness                               | Ensures consistent indentation (no mixed spaces/tabs) within command sections.                                                              |                                                         |
 | `ConciseInput`              | Style                                                       | Ensures concise input assignments are used (implicit binding when available).                                                               |                                                         |
 | `ContainerUri`              | Clarity, Portability                                        | Ensures that values for the `container` key within `runtime`/`requirements` sections are well-formed.                                       |                                                         |
 | `DeclarationName`           | Naming, Style, Clarity                                      | Ensures declaration names do not redundantly include their type name.                                                                       |                                                         |
