+++ crates/wdl-lint/src/rules.rs
@@ -2,7 +2,6 @@
 
 mod bash_set_syntax;
 mod call_input_keyword;
-mod command_section_indentation;
 mod concise_input;
 mod container_uri;
 mod declaration_name;
@@ -40,7 +39,6 @@ mod unused_doc_comments;
 
 pub use bash_set_syntax::*;
 pub use call_input_keyword::*;
-pub use command_section_indentation::*;
 pub use concise_input::*;
 pub use container_uri::*;
 pub use declaration_name::*;
