+++ tests/cli/explain/tag/All-JSON/stdout
@@ -1,1 +1,1 @@
-[{"name":"All","applicable_lints":["BashSetSyntax","CallInputKeyword","CommandSectionIndentation","ConciseInput","ContainerUri","DeclarationName","DenyGlobStar","DeprecatedObject","DeprecatedPlaceholder","DeprecatedRuntimeSection","DescriptionLength","DocCommentTabs","DocMetaStrings","DoubleQuotes","EmptyDocComment","EmptyOutputs","ExpectedRuntimeKeys","HereDocCommands","HostPathLiterals","ImportPlacement","InlineInstall","InputName","MatchingOutputMeta","MetaDescription","MetaSections","OutputName","ParameterDescription","ParameterMetaMatched","PascalCase","RedundantNone","RequirementsSection","RuntimeSection","SectionOrdering","ShellCheck","SnakeCase","TodoComment","UnusedDocComments"]},{"name":"Clarity","applicable_lints":["CommandSectionIndentation","ContainerUri","DeclarationName","DenyGlobStar","DocCommentTabs","DoubleQuotes","EmptyDocComment","HereDocCommands","ImportPlacement","MetaSections","PascalCase","SnakeCase"]},{"name":"Completeness","applicable_lints":["EmptyOutputs","ExpectedRuntimeKeys","MatchingOutputMeta","MetaDescription","MetaSections","ParameterDescription","ParameterMetaMatched","RequirementsSection","RuntimeSection"]},{"name":"Correctness","applicable_lints":["BashSetSyntax","CommandSectionIndentation","DenyGlobStar","HereDocCommands","InlineInstall","ShellCheck"]},{"name":"Deprecated","applicable_lints":["CallInputKeyword","DeprecatedObject","DeprecatedPlaceholder","DeprecatedRuntimeSection","ExpectedRuntimeKeys"]},{"name":"Documentation","applicable_lints":["EmptyDocComment","MatchingOutputMeta","MetaDescription","MetaSections","ParameterDescription","ParameterMetaMatched","UnusedDocComments"]},{"name":"Naming","applicable_lints":["DeclarationName","InputName","OutputName","PascalCase","SnakeCase"]},{"name":"Performance","applicable_lints":["InlineInstall"]},{"name":"Portability","applicable_lints":["ContainerUri","HostPathLiterals","InlineInstall","RequirementsSection","RuntimeSection"]},{"name":"Sorting","applicable_lints":["ParameterMetaMatched","SectionOrdering"]},{"name":"Spacing","applicable_lints":["CommandSectionIndentation"]},{"name":"SprocketCompatibility","applicable_lints":["DescriptionLength","DocMetaStrings","MatchingOutputMeta","MetaDescription","ParameterMetaMatched"]},{"name":"Style","applicable_lints":["CallInputKeyword","ConciseInput","DeclarationName","DocCommentTabs","DoubleQuotes","InputName","OutputName","PascalCase","RedundantNone","SectionOrdering","SnakeCase","TodoComment"]}]
+[{"name":"All","applicable_lints":["BashSetSyntax","CallInputKeyword","ConciseInput","ContainerUri","DeclarationName","DenyGlobStar","DeprecatedObject","DeprecatedPlaceholder","DeprecatedRuntimeSection","DescriptionLength","DocCommentTabs","DocMetaStrings","DoubleQuotes","EmptyDocComment","EmptyOutputs","ExpectedRuntimeKeys","HereDocCommands","HostPathLiterals","ImportPlacement","InlineInstall","InputName","MatchingOutputMeta","MetaDescription","MetaSections","OutputName","ParameterDescription","ParameterMetaMatched","PascalCase","RedundantNone","RequirementsSection","RuntimeSection","SectionOrdering","ShellCheck","SnakeCase","TodoComment","UnusedDocComments"]},{"name":"Clarity","applicable_lints":["ContainerUri","DeclarationName","DenyGlobStar","DocCommentTabs","DoubleQuotes","EmptyDocComment","HereDocCommands","ImportPlacement","MetaSections","PascalCase","SnakeCase"]},{"name":"Completeness","applicable_lints":["EmptyOutputs","ExpectedRuntimeKeys","MatchingOutputMeta","MetaDescription","MetaSections","ParameterDescription","ParameterMetaMatched","RequirementsSection","RuntimeSection"]},{"name":"Correctness","applicable_lints":["BashSetSyntax","DenyGlobStar","HereDocCommands","InlineInstall","ShellCheck"]},{"name":"Deprecated","applicable_lints":["CallInputKeyword","DeprecatedObject","DeprecatedPlaceholder","DeprecatedRuntimeSection","ExpectedRuntimeKeys"]},{"name":"Documentation","applicable_lints":["EmptyDocComment","MatchingOutputMeta","MetaDescription","MetaSections","ParameterDescription","ParameterMetaMatched","UnusedDocComments"]},{"name":"Naming","applicable_lints":["DeclarationName","InputName","OutputName","PascalCase","SnakeCase"]},{"name":"Performance","applicable_lints":["InlineInstall"]},{"name":"Portability","applicable_lints":["ContainerUri","HostPathLiterals","InlineInstall","RequirementsSection","RuntimeSection"]},{"name":"Sorting","applicable_lints":["ParameterMetaMatched","SectionOrdering"]},{"name":"Spacing","applicable_lints":[]},{"name":"SprocketCompatibility","applicable_lints":["DescriptionLength","DocMetaStrings","MatchingOutputMeta","MetaDescription","ParameterMetaMatched"]},{"name":"Style","applicable_lints":["CallInputKeyword","ConciseInput","DeclarationName","DocCommentTabs","DoubleQuotes","InputName","OutputName","PascalCase","RedundantNone","SectionOrdering","SnakeCase","TodoComment"]}]
