+++ tests/cli/explain/tag/Clarity-JSON/stdout
@@ -1,1 +1,1 @@
-{"name":"Clarity","applicable_lints":["CommandSectionIndentation","ContainerUri","DeclarationName","DenyGlobStar","DocCommentTabs","DoubleQuotes","EmptyDocComment","HereDocCommands","ImportPlacement","MetaSections","PascalCase","SnakeCase"]}
+{"name":"Clarity","applicable_lints":["ContainerUri","DeclarationName","DenyGlobStar","DocCommentTabs","DoubleQuotes","EmptyDocComment","HereDocCommands","ImportPlacement","MetaSections","PascalCase","SnakeCase"]}
