+++ tests/cli/explain/tag/Clarity/stdout
@@ -1,5 +1,4 @@
 Rules with the tag `Clarity`:
-  - CommandSectionIndentation
   - ContainerUri
   - DeclarationName
   - DenyGlobStar
